crawler/requirements.txt
# **************************************************************
# MAINTENANCE BOX
# 10.10: JC - Add Selenium package for Azure build time
# **************************************************************

# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob==12.5.0
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
crawler==0.0.2
idna==2.10
multidict==4.7.6
panda==0.3.1
requests==2.24.0
selenium==3.141.0
soupsieve==2.0.1
urllib3==1.25.10
yarl==1.6.0